Bumps core for sidekiq DB pool configuration
This will now default to 25 connections per server. This is inline with
the default sidekiq concurrency of 25 workers per server.

This is configured via code, to ensure it's only loaded in the context
of a sidekiq server. Also, dicking around with a hybrid of
`database.yml` and `DATABASE_URL` would be sad and it's Friday.