Lock down `devise_invitable` to prevent users inviting others.
This is accessible via `/users/invitations/new` but not linked anywhere.
Still, we should make sure it's impossible. It'd be nice to remove it,
but that would involve adding a new override controller which seems
a bit heavyweight.

As it is, we let users invite a maximum of 0 users (as the docs suggest)
and also modify the flash message display to show the "you've got no
invites" flash message if anyone ever tries (they won't).