Amend 'Add trading name' forms to handle pre-existing record.
We've got a situation where there are a number of incomplete records in
the database. This causes a number of interesting phenomena. In this
instance, we look to see if there's an incomplete copy of our requested
firm first and use it if it's there. Otherwise we'd need to delete it,
because we'd have a dupe.

Our forms weren't written to expect this, so when the 'new' form was
passed an existing record it created an unusual URL. To avoid this, we
point the form to the edit action if the record is persisted, and the
create action if it's new.