Fix suspected cause of empty adviser requests.
We've got this weird occasional error where the client requests
`/lookup/advisers/.json` even though there should be a guard against
this. Can't reproduce though.

The adviser number and the check run against objects that might
plausibly be different, so I've made them exactly the same. Should
probably be like this anyway.

Also added a comment to caveat something in the existing implementation.