Merge `mas-rad_core` into `rad`
More than 85% of `mas-rad_core` is used by `rad` only.
A small percentage of the remaining 15% is "shared" between
`rad_consumer` and `rad`: we believe that this is not enough to
justify the `mas-rad_core` existence in the first instance.

This commit adds back what belongs to `rad`, and the same will
be done for `rad_consumer` in order to create a better separation
of concerns, if it needs to be.