Fix flaky snapshot test
The snapshot test was relying on the `match` matcher to check
for inclusion _and_ order of the queried entities (i.e. firms),
but somehow we don't need to test the _order_ as well, since the
query is done by using a postcode, which in turns can give a bunch
of results, no matter what's the actual "order" (since it's not defined!)

So we use `match_array` now to check for inclusion only.

It also turns out that ActiveRecord does not ensure an order,
unless explicitly set. More info https://github.com/rails/rails/issues/13743

Co-Authored-By: Marco Bagnasco <marco.mkrs@gmail.com>