Bugfix: consistently upcase reference number when building adviser
Ensures adviser is built with upcased reference number, for consistent
comparisons with the lookup table advisers.

One scenario this bug occurs in is if a user tries to add an adviser to
a firm with Javascript disabled in the browser. The JS that upcases the
user input isn't available and so the server returns an error because it
can't find a matching adviser in the lookup data.