Wrap principal and user deletion with transaction
When a principal gets removed, the associated firm and remaining
associations are removed, this is because there's a
`dependant: :destroy` association between them.

The user instead, while it's associated to the principal,
needs to be removed with a separate query.

Since I am not sure about the consequences of adding
`dependant: :destroy` to the user<->principal association,
the safest approach is to wrap the two statements
to delete the principal _and_ the user in a transaction block.