Fix counters not being updated
To keep counters up to date, you need to update _all_ advisers'
records for a firm, specifically when:

- An adviser is added/updated/deleted
- An office added/updated/deleted

To do so, we need to know about the firm, but we are enqueuing
a job by passing passing the object `id` only, we need
to retrieve the `firm_id` from somewhere.

If the object still have a db record, retrieve it from there,
otherwise rely on an additional `firm_id` argument.