Notify admins when new firm is registered
When a principal registers a new firm in the directory, we want
to notify the admins of the directory of the new firm added so
they can immediately check the firm information.

In order to do that, we are sending an email to the RAD admins
account containing a link to the new firm in the directory admin
panel.