Fix tests
Lots of tests are breaking because they use a factory to create a
firm or principal. Because the validation of a principal/firm relies
on calling the FcaApi to check the existence of the fca number, validation
fails on create. Stubbing any instance of the FcaApi::Request class handles
provides a way round calling the api.