Fix tests
Lots of tests are breaking because they use a factory to create a
firm or principal. Because the validation of a principal/firm relies
on calling the FcaApi to check the existence of the fca number, validation
fails on create. Create a shared_context to stub the FcaApi::Request and
Response.