Add timestamp to record approval of Firms
We want to be able to track if a Firm is approved and when that approval
happened.

Using a datetime field that can be NULL (and will be NULL by default),
we cover both requirements.

We will assume all the existent firms as already approved, hence the
backfill with the current timestamp.
Only the new firms after this migration will miss this value until its
set on firm approval.