Remove logic to reindex old firms
This is a leftover.

There is no need to re-index firms in this way when moving advisers from one firm to another. Detailed explanation:

Before the call to re-index was a _syncronous_ call to a service that was taking the model instance as argument (which still keeps the _old_ `firm_id` when `after_commit` is triggered: this is because you would need to do another trip to the database after commit to receive the new instance.

Since we use an _asyncronous_ job now
to (re)index and it doesn't take an object as arguments like before, but the id of the adviser instead, by the time the worker picks up the job and queries the db, the new `firm_id` is definitely there.