Housekeeping
- Add environment variabless to docker.
- Fix broken specs using include_context and forcing callbacks.
- Remove feature that tests the fca number look up because we dont
do that any more, we use the fca api to check the fca number.

A previous pr to improve the algolia feature specs set a global
flag to force callbacks. This has the effect of breaking tests which
use a Principal because Principal calls the fca api using an after_commit
hook. This commit reverses the global flag and sets it for
the specific tests which require the after_commit action to run.