Cover directory updates with integration tests
Current integration tests cover the management of
Principals, Firms, Advisers and Offices from both the Principals and
Admins user flows.

The test coverage is missing a key part of RAD business logic though:
"What and when gets pushed/removed to the RAD Directory?"

The main purpose of RAD, besides allow Principals/Admins to maintain
firms advisers and offices, is to provide that data to the public
consuming Retirement Adviser Directory. And that should be covered
with integration tests.

The retirement advisor directory publicly accessible information gets
pushed from this app to Algolia. Containing:
- One index for firm Advisers.
- One index for firm Offices.

Being able to inspect what the index contains at the end of user flows
is quite tricky, and can have different approaches:

(TLDR: Jump to the fifth Approach - We used an Algolia Test Double)

* First approach: Use real Algolia cloud stored test indexes.

We have shared Algolia test indexes running exactly as in production.
The problem is that this data is shared between different test/runs and
data from one test can pollute the expectations of other tests.

On top of the test isolation issue, we would have a real dependency on
the ALGOLIA API KEY for testing and we would need to hit
Algolia service multiple times for each feature test run.

* Second approach: Set expectations over Webmock/VCR cassetes.

This would cover when we want to check that "this object gets pushed to
Algolia" or "this object gets removed from Algolia", but it would imply
a quite big number of cassettes and also doesn't make easy to just check:
"Algolia Index contains X but not Y" expectations at the end of an user
flow.

* Third approach: Use Algolia provided testing Mocks.

Algolia gem provices Webmock stubs collection for testing purposes.
Problem is that those stubs return fixed data, with fixed ID and
information. If we want to test what do we actually push/delete from
our indexes, this approach wouldn't allow us, as indexes will always
return the same unrelated dummy data.

* Fourth approach: Just set expectations over calls on Algolia Indexes.

Classic "I expect Algolia to receive a call over add element to the
index".

This stubbing works great for Unit testing, when you isolate bits of
business logic and just expect "X to be called", but we are doing
end to end feature testing, and we mainly want to check "What Algolia
contains after this user flow happen".
In order to get anything similar
to that setting call expectations over Algolia Index stubs, we would
need to know too much about the implementation and define lots of
stubs, making tests quite brittled and coupled with the implementation.

* Fifth approach: Replace real Algolia for a local test double.
We don't want to be stubbing all the Algolia calls, we don't want to
use real Algolia service with network dependency and lack of test env
isolation... what do we have left?
We can just replace Algolia all together with a fake test directory
providing the same interface as Algolia but just storing the indexe
into an array.

This allow us to:
- Not have any network/api dependency.
- Test end to end what indexes contain at the end of an user flow.
- Minimal use of Stubs (we only stub the calls for instantiating a
  new Algolia::Index object).
- Reduce the coupling with implementation details.
  As a single spec context file contains all the Algolia::Index stub
  and helper methods to query the index from tests.
  If something happens to change in real Algolia gem interface/service,
  we basically have a single point to maintain, instead having
  implementation details polluting all our test files.

After trying/examining all the options, went for the fifth/last
approach, as allows us to test what we want to test with minimum
pain.

(cherry picked from commit 6b5d09a59a12a1c03d63097194405e347aa8201e)