Add geolocation data where missing on Office records
Seven office records are missing geolocation data. It's unclear why this
is, however all of them appear to have been added to the database at the
same time - Mon 12th Oct 2015.

Call the geocode method on each record and persist the changes.

Context for this change: a bug reported in production whereby an office
added to Firm ID 1269 (Bluesky Wealth Management Limited) is not
appearing in the RAD consumer frontend.