Create the associate firm
Prior to this change, the creation of the associate firm was
triggered by an after_commit hook when a principal was created.

Because we now get the firm's name from the api, the after_commit
approach is no longer appropriate. Instead, we grab the name from
the api call and explicitly create the principal's associate firm.

Due to the change in approach to creating a firm, alot of tests
broke because they were reliant on the automatic creation of a firm
once a principal was created. Alot of the changes in this commit
are to fix these tests by explicitly creating the principal's
associate firm.

- Get the name associated with the Firm Reference Number(FRN)
from the FCA API.
- Use the name to create the associate firm along with the FRN.
- Amend the FcaApi::Request to facilitate querying the information
the FCA API returns.
- Housekeeping - log results of FCA API call.
- Fix a load of tests by amending the principal factory to build
a firm, rather than a lookup firm.