Increase Sidekiq concurrency
We currently have Sidekiq's database connection configured with a pool
size of 25 (see ./config/initializers/sidekiq.rb'), however we have the
number of concurrent workers set to 10.

Sidekiq's own documentation recommends the two numbers be the same. It's
not entirely clear from the commit history why there's such a large
disparity between these two numbers in the current config.

Increase concurrency to 20, for the following reasons:

* Greater utilisation of the available connection pool.
* An upcoming feature related to status validation of firms via the FCA
  API is expected to require at least 20 threads to efficiently process
  all firms while dealing with an API rate limit.