Fix Algolia adviser spec
This spec has started breaking following the FactoryGirl > FactoryBot
changes, presumably because of a change in how create_list works. The
error relates to a duplicate primary key when creating a set of advisers
in the test setup.

Change the spec so that we no longer hardcode the id of one of these
advisers.