Change Jenkins test suite invocation
Update the Jenkins test script so that we run feature specs separately
from unit tests.

Why?

The RAD test suite currently experiences intermittent failures.

Some of these errors appear to be due to some leaking state from the
feature specs, causing intermittent errors when the test suite is run as
a whole. The errors relate to some kind of database transaction issue,
seemingly originating from the FCA file import code.

Until we isolate the source of this, do two distinct runs to address
some of the intermittent failures and get us closer to a green build.