Change how we link to modernizr lib
We're currently using Rails.application.assets to access the modernizr
JS library, which is included with dough. It's unclear why this method
was chosen. As of the recent Rails 5 upgrade, the value of the
configuration key in production is nil, which leads to errors.

Replace this with a Sprockets link declaration in the manifest. This
will resolve to the dough vendor/assets/non_bower_components folder
(because vendor/assets in an engine is an asset lookup path by
convention) and will precompile modernizr making it available as a
public asset. This is then linked to via a standard
javascript_include_tag in the layout.