Change Firm validation behaviour
There are several attributes on the Firm model that are not populated
until some time after initial Firm creation. That is, a firm can
register with their FCA number alone and further details are provided as
part of their "onboarding" process.

Validations on these additional details make no distinction between
create and update. The codebase gets around this by skipping validations
when we create a new firm, using a similar hack in the firm factory.

Rails has in-built functionality to deal with this scenario - validation
logic can be set to run on update-only. Apply this change to the Firm
model and improve the validation specs.