Firm factory change - emphasise not_onboarded trait
- Rename the 'invalid' trait to 'not_onboarded'.
- Change the factory that uses it such that its primary name is
  'not_onboarded_firm' rather than that being an alias.

A small change, but the intention here is to encourage the use of
'onboarding' to describe the process of a firm providing all the details
required for it to be eventually considered publishable on the
directory. These are the Firm attributes which are validated on update
rather than create. It also de-emphasises the use of the word
'invalid' for these firms, as a Firm not having those attributes is a
valid state within the system.