Replace 'registered' with 'onboarded' when talking about Firms
The word 'registered' is used at the Firm level to describe the fact
that the firm has provided a set of details that are required before it
can be published. This is vague, potentially confusing, and overloads
usage of the word 'registered' within this codebase. Additionally,
'onboarded' is used in a couple of places to allude to approximately the
same concept.

Instead, describe this state as 'onboarded' and use this vocabulary
consistently. As such:

- Registration is the process of creating a principal account and an
  associated firm, saved with its registered name and an FCA number. It
  is also the process of adding other firm types to an existing
  principal account.
- Onboarding is the process of providing further details about the firm
  such that it can eventually be published on the relevant directory.