Fix firm deletion
Firm deletion is currently broken for some firms. This is because there
are duplicate firm IDs in the inactive_firms table. What this means is
that even though the inactive_firm relation on Firm is dependent
destroy, a foreign key constraint gets triggered because a related
record still exists in inactive_firms.

Address with the following:

- Remove the foreign key constraint. The inactive_firms table is used
  for reporting. It's more important that the business logic for Firm
  deletion works correctly than it is for referential integrity to exist
  between these two tables.
- Add a unique index on firm_id in the inactive_firms table.