Consistently downcase Principal email addresses
All emails stored in the Users table are lower case. The codebase also
stores email addresses in the Principals table. There doesn't appear to
be an obvious reason for recording this information twice, and it may
well be a design flaw that needs correcting.

In the meantime, however, we have an issue whereby the Principal emails
are stored in whatever case the user provides. This causes issues when
retrieving records, in that we have to be aware of case-sensitivity in
any queries.

Address this as follows:

- Add a callback to Principal to ensure the email is downcased before
  the object is validated and persisted.
- Add a migration to downcase all existing Principal emails.
- Add logic to the NewPrincipalForm that consistently downcases the
  email fields. This may seem redundant given the callback on the model,
  however the data in this form is also used to check for existing
  principals by email address in cases where a principal is signing up
  their firm for more than one directory.