Add a TODO note to AlgoliaIndex::TestSeeds
This file was part of the process of using an Algolia index in the test
suite. That approach has since been replaced by either stubbing the API
calls or using AlgoliaIndex::Fake.

This project is being handed over to a new development team today - add
a note to review this file at a later time.