Escape html in a name matcher spec.
Faker can generate names that have apostrophes in them and they cause an intermittent test failures.

Christopher Wall' ace doesn’t match with Christopher Wall&#39; ace

Rather than gsubbing out the special chars in the factory, opted to escape the string we’re matching against because IRL there might be names with special chars so better not to exclude the possibility in tests.