Fix spec non-determinism
This was tricky to trace but when the specs run with specific seeds
(thus specifically ordered) there are occassions where the
`Algolia::Index` is already memoized as non-fake. Any subsequent specs
would fail thereafter. This could be traced back to this missing,
expected context to be around for the feature I have modified in this
commit.

An example failing seed would be `49365`.