Correctly include `geocoded` scope
Where this concern was included the chained scope was not correctly
included in the resultant SQL query. For example:

`adviser.firm.offices.geocoded.count`

would yield the geocoded offices but without being scoped to `firm`
hence the total number of ALL geocoded offices was returned. This
applied in any chained scope where `geocoded` was used. I added the
`TestSeedsSpec` back as it was the only test that this problem would
manifest.